Fix MMIO byte width and cmp+equality branch translation bugs
Three translator bugs fixed:
1. mmio_read/mmio_write used u16 pointers generating lw/sw (word ops),
   but COR24 I/O registers are byte-addressed per hardware Verilog.
   Changed to u8 pointers → mov.b → lb/sb. All 12 demos regenerated.
2. mov.b (byte mode) was ignored by translate_mov — now generates
   lb/lbu/sb instead of lw/sw when byte_mode is set.
3. cmp+jeq/jne used clu (unsigned less-than) but equality branches
   need ceq. Added look-ahead to select ceq vs clu based on following
   branch type. Also fixed cmp-immediate clobbering temp register
   without push/pop save/restore.

Added demo_echo_v2: Rust logic (to_upper, handle_rx) with minimal
asm!() for COR24 interrupt plumbing only. Passes end-to-end test.

+++ rust-to-cor24/demos/demo_nested/src/lib.rs
@@ -11,13 +11,13 @@ const UART_DATA: u16 = 0xFF01;
 #[inline(never)]
 #[no_mangle]
 pub unsafe fn mmio_write(addr: u16, val: u16) {
-    core::ptr::write_volatile(addr as *mut u16, val);
+    core::ptr::write_volatile(addr as *mut u8, val as u8);
 }
 
 #[inline(never)]
 #[no_mangle]
 pub unsafe fn mmio_read(addr: u16) -> u16 {
-    core::ptr::read_volatile(addr as *const u16)
+    core::ptr::read_volatile(addr as *const u8) as u16
 }
 
 #[inline(never)]
